Bug 1122840 - Split storage into a separate framework.
This should improve build times and space efficiency.

Note that we add a modulemap to Storage. This allows us to import
sqlite3.h without the dreaded "include of non-modular header inside
framework module" error.

We set

        CLANG_ALLOW_NON_MODULAR_INCLUDES_IN_FRAMEWORK_MODULES = YES;

but it doesn't do anything.

This commit makes a few things immutable, reworks a few initializers as
a result, and otherwise slightly alters accessibility to make
cross-module usage possible.