Bug 1135814 - Content should not be under the bottom toolbar
Added constraints so that when a tab is initially loaded the webView's frame is
limited to the area between both toolbars.  This prevents pages that use the
full screen and disable scrolling, from displaying content beneath the bottom
toolbar.  The view constraints are changed to use the full screen and toolbars
can be docked away, whenever there is content beyond the bottom toolbar that
requires scrolling to reach.