Bug 1139051 - Post: Only run live tests if signedInUser.json includes an email address.
It's not easy to include a conditional resource in the test bundle, so
we always include an empty signedInUser.json.  A developer wishing to
test should overwrite the empty file with their own.

Since there's no way to recover the password from signedInUser.json (or
its equivalent on any platform!), this test code assumes that the test
account's password is identical to its email address.