Bug 1151387 - Part 1: Hide location view when editing
While this has not *visible* effect, it has effect on accessibility
which then ignores the view (which is the correct behavior).

Otherwise we would be able to access location view elements (like url
label, reader button, reload button, lock image etc.) even when editing,
i.e. when they are not visible.