Bug 1129505 - Part 1: Persist FxAState as JSON.
This paves the way to persist the JSON independently from the account.
The change is motivated because it's easy to /serialize/ a Swift
dictionary to a String (use JSON), but it's not easy to /deserialize/ a
Swift dictionary from a String (JSON only gives you back wrapped
dictionaries).  So deserialization requires using JSON; hence this
commit.