Bug 1151504 - Location view not visited by Switch Control
This also slightly improved things for VoiceOver users as this is a hint
the location label can be activated and something happens.

Best long term solution would be to change the element from a label to
a text field, as that best fits the semantics of what the element
actually does (displays a text value and when tapped, allows editing
it). VoiceOver should then read "text field" for it, and Switch Control
should automatically visit it without any special traits set.