Bug 1160781: Accessibility escape gesture audit
Previously the gesture did not work correcty where we are in the home
panel or search controller and VoiceOver cursor is not in the url bar -
in this situation, it did not dismiss the home/search panel and instead
did "go back" in the current tab (i.e. that was wrong).

Now the gesture behaves this way:
- in (non-inline) home panel and search controller: we dismiss them
- when displaying webpage: go back if possible