Introduce a subquery to avoid doing extensive grouping of visits.
This change essentially reintroduces the original query -- one row per
history URL.

We then wrap this in another query that groups by domain, summing
frecency. As before, we then also wrap in another query that joins
against the icon table.

This avoids the need to do 20x the work in the in-memory b-tree, which
saves a lot of query time.

Additionally, this eliminates a dead row when aggregating against an
empty result set, which was causing a crash in the row factory when
displaying top sites for an empty profile.