Bug 1170707 - Synchronize FirefoxAccount.advance() invocations. r=etoop,st3fan f=rnewman
We have a classic data race here, made interesting because one of the
initialization tokens from the fxa-content-server is single-use.  What
happens is that multiple parts of the code advance() the (single)
FirefoxAccount.  They all start with the same initial datum, racing to
use the single-use token.  One connects first, remotely invalidating the
token, and continues to do slow crypto things.  Others lose the
connection race and see a 401/110 from the fxa-auth-server since the
token is now invalid.  The latter are displayed in the UI, even though
the first eventually succeeds.

This patch introduces a shared deferred and a lock for updating that
deferred.  I specifically want a reviewer other than (or in addition to)
rnewman for this, since he sketched the solution and I want to ensure I
communicated it well enough for a disinterested third party to
understand.