Bug META - Rework SyncStateMachine.
This moves from a ternary Maybe<SyncState> where a failureValue may
possibly be a RecoverableSyncState or a terminal failure, to a simpler
Maybe<SyncState> where a failureValue is always a terminal failure.  It
makes SyncStateMachine a concrete class, and extracts the state machine
advancing logic out of the SyncState instances themselves.