Bug META - React to remote meta/global changes.
If we can parse the remote meta/global, we can proceed.

I haven't handled a situation when we need to upload a fresh
meta/global.  The only situation where this might happen is when we see
a meta/global with an engine record corresponding to a locally
implemented engine, but with an incorrect version.  If the version is
too new, we can't do anything but ignore the entry.  If the version is
too old, we could wipe and update meta/global.  The complexity to
achieve the latter is high; let's leave it for now.