Fix update of private tabs when switching off private browsing.
When switching off private browsing with the "Close Private Tabs" option
on, tabs were closed and notified the delegate. But at this stage,
notification should not be emitted at it conflicts with the view change
state.

This patch inhibits the call of the delegate when quitting private
browsing with the "Close Private Tabs" option on.