Bug 1228611 - Part 5: forcibly shut down the profile in applicationWillTerminate. Log open profile file descriptors on background in debug builds.
Just dropping all references to the BVC and profile isn't enough --
something is still holding a reference. So we forcibly close during
applicationWillTerminate.

In the newly introduced background() call we log our open file
descriptors. If we see more than one reference to a DB, we might be in
trouble.