Bug 1230288 - Forcibly shut down the profile in applicationWillTerminate, rather than closing the database connection when backgrounding.
This is a follow-up from Bug 1228611.

Just dropping all references to the BVC and profile isn't enough --
something is still holding a reference. So we forcibly close during
applicationWillTerminate.

In the newly introduced background() call we log our open file
descriptors. If we see more than one reference to a DB, we might be in
trouble.

The big change here is that we don't routinely close the database
connection when backgrounding. This avoids some flapping. Because we
close the connection in applicationWillTerminate, this also makes sure
that we close any connection we open _in the background_ due to a
delayed sync.

We might consider doing the latter regardless.