Bug 1230623 - Forcibly shut down the profile in applicationWillTerminate.
This is a follow-up from Bug 1228611.

Just dropping all references to the BVC and profile isn't enough --
something is still holding a reference. So we forcibly close during
applicationWillTerminate.

Usually there will be nothing to do, because before Bug 1230288 we close
our connections when we're backgrounded. But a delayed sync (e.g., after
a login change) can reopen one, so this change avoids the possibility of
our process being terminated with database connections open.