Bug 1247482 - Part 2: use withExtendedLifetime to protect pragma queries.
This is the place we crash on my device with whole module optimization enabled.

The compiler appears to erase the constant strings or the retains needed to
protect non-constant query strings; by the time we get to `pragma` we die with
an EXC_BAD_ACCESS due to a null `String`, which should be impossible.

Adding `withExtendedLifetime` makes sure the query strings stick around.