Bug 1243847 - Part 2: Import en search plugins locally
Since the en search engines are part of the main Android repository and
not committed to a separate l10n repo, the import script won't detect
the en search plugins. This maintains a local copy of the en plugins
directory and copies it into SearchPlugins each time the import script
is run.