Bug 1262154 - URLBarView Version 2
Originally this bug was to create a URLBarView that supported a flexible
layout where we move the home panel icons into it while in landscape.
Now it has become more of a larger refactor including fixing janky
animations, rendering the curve background properly, and increasing the
responsiveness of the bar by allow touch interactions during animations.
Also includes a general clean up of how we manufacture the toolbar items
inside the URLBarView such as the tabs button and basic
back/forward/share etc buttons.

The current state of this branch is in a partially working state
containing a new URLBarView named URLBarView_V2 that adheres to an 1-1
mapping of the existing URLBarView API. Some features that are missing:

* Themeing. Currently WIP
* Text squish. WIP in removing the squish but allowing instant editing
  of text field.
* Accessibiliy review. Maybe need to check that accessiblity is valid
  for new views.