Bug 1267241 – Align Copy Link label with default padding.
This is complicated by the fact that we don't know what the padding is until the framework tells us, and we don't want to move the other buttons.

Solution: we offset the copy link text to the default padding position, then move the image relative to that.

We don't actually accept the default padding.

This isn't enough however.