Bug 1273963 – Inform Webkit of user response to prompt/confirm.
The JSPromptAlertController was doing calling the completionHandler (i.e. telling webkit) _before_ the alert action handler was being called, so the web content was getting the wrong values – which may have been different to those that the user had selected.

Since the alert view was being dismissed before the action handler was being called, we can simplify the code by calling the completion handlers from within the action handlers themselves.

This means we can no longer need track the state as properties, so we can get rid of the mutating keyword, and the change vars to lets in BVC.