Bug 1211201 – Accessing the URL bar can take more scrolling than expected
Previously, a touch target existed to let the user show the URL bar
when it was hidden, but the feature was not at all obvious. Now it is
represented by a much smaller URL bar that users can tap on to bring
the main bar back, without obscuring much of the page.