Bug 1304035 - correctly detect whether current page is web page
Recent changes amended the test to determine whether a current scheme is a web page from isHttpOrHttps || isDataURI to isHttpOrHttps where isDataURi, which could never be satisfied as the scheme cannot be both http && data at the same time.
Returning the test to OR instead of AND reenables the share button (and the reload button which was also disabled).