Bug 1304716 - The selected row remains highlighted (e.g. 'Logins') when Canceling the Passcode request prompt
Bugzilla bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1272658

https://bugzilla.mozilla.org/show_bug.cgi?id=1304716

The reason why this was happening on iPad (And also on iPhone 6/7+ on landspace) and not on iPhones is because when a modal is dismissed it will call “viewWillAppear:” on “SettingsTableViewController” which calls “reloadData” on the tableView.

But on iPads this is not true since the modal is being called on a FormSheet factor, therefore the presenting controller's view never disappears, so “viewWillAppear:” is not called after the dismissal.