Bug 1272231 - The logins are displayed as deselected after screen lock
Bugzilla bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1272231

The issue was happening because the `loginSectionsDidUpdate` was being called after the user typed the password and the screen refreshed, and this method calls `reloadData` which cleans the selections.

The fix is done by looping on all selected indexPaths stored on `loginSelectionController` and selecting them again.

Some `self` references were also removed since they are not required, as described on the style guide: https://github.com/raywenderlich/swift-style-guide#use-of-self