Bug 1321576 - Removes isUserZoom and caching of zoomScale
One issue with assigning the zoomScale inside scrollViewDidZoom is that
the WKWebView will invoke this method when it changes it's own scale
such as requesting a page in Desktop Mode. Removing the handling of this
and keeping the update minimum scaling logic on rotation seems to
resolve this and the frictionless page issue that was caused by
incorrect page scaling.