Review comments: Don't show toast when top tabs are visible; extract string constants.
I didn't extract additional strings that were nearby, since it looks
like this file isn't really using the Strings.* style yet.

I elected to duplicate the "Switch" button label since it may be that
some locales want to use different verbs for regular and private tabs;
or we may want to color or border the button differently to emphasize
that the new tab is private in future.  In that case, it'll be harder
if we try to unify the button label and styling now.