Remove newlines from JSON.rawString and pass in URL string when injecting into the FxA webview
Fixed SwiftyJSON's use of rawString with no options which causes
newlines characters to be included into the rawString. This was causing
the script we inject into the FxA web view to be invalid.

Additionally, the explicitly defined URL property was being passed as a
normal String in Swift 2.3 but in Swift 3 is looks like even if a
property is defined as explicitly unwrapped, it will still print
Optional("value") when being injected into a string. Using `absoluteURL`
works around this.