Fixed async result of updating malformed logins URL
We previously we blocking on the update call by calling .value on it
which we removed because it didn't make sense. In turn this caused our
test to fail since the test calls .value on the deferred to force block
it but the update is async. To correct the flow, we should be chaining
the deferred off of the result of the update call.