Properly reset data on disk when creating a new profile
Flips the boolean check which was incorrect before from the Swift
migration. Additionally, since DBs are now not lazy, we need to force
clear the profile when opening in a test because they now have a
presence on disk immediately where as before they never did until the DB
was used.