Bug 1341626 - ‘Go to copy link' toast is displayed for 1 second and it gets dismissed
Bug link:https://bugzilla.mozilla.org/show_bug.cgi?id=1341626

This was happening because `showHomePanelController` is called multiple times to render the home screen, and when this happens the subview hierarchy is changed.

The fix just checks if the toast is being displayed, and brings it to front if necessary.