Bug 1345409 - Ensure that we always close the selected tab last when closing all private tabs
In top tabs, trying to close the selected tab causes a new tab to be chosen, but by the time the code comes to render it, it may no longer exist.
Ensuring we close the selected tab last means that the new selected tab will come from the non-private list and no crashing happens.