Part 2 - Attach a DB called metadata.db to browser.db
If the DB doesn't exist it will be created

This cannot be done inside a transaction - this is a quirk of the `ATTACH` command, and because we reference the metadata table before we insert anything into it, through top sites and highlights, we cannot wait to create or attach the DB until we start to populate its tables, so we have to do it here.

Also, as `logins.db` also implements `BrowserDB`, we cannot do it during the init of `browser.db` unless we want to attach `logins.db` too. Perhaps a follow up should be filed to separate Logins and Browser