Pre: Creates new framework/target for Telemetry
Adds a new framework to contain all of our Telemetry instrumentation
code and separates the target-specific ping events to their respected
targets. For example, CorePing/ClientTelemetry now live with the Client
target and SyncTelemetry lives in the Sync module but pull in Telemetry
as a dependency for the definitions and instrumentation.