Bug 1354935 - Part 1: correct SwiftyJSON type checks. r?sleroux
SwiftyJSON doesn't tightly couple its type enum with values, so it's technically
possible for a .string to have a value that returns nil when failably cast to a
string.

Additionally, some SwiftyJSON error flows set type to .unknown.

We were mishandling both of these steps: assuming that if .type == .string then
.string would always return a string, and that .unknown meant a null field.

This commit fixes the behavior to match the pre-SwiftyJSON state: .isNull() is
true when the value is an explicit null, and .isString() is true when .string is
non-nil and the type is .string.