Bug 1392317: settings footer text longer than 2 rows is cut off (#3087)
The TP settings screen has longer table view section footer labels that what we had previously,
and we didn't have code to handle growing the footer row height dynamically.
This is now fixed, the footer will either be 44 points, or grow larger if the text is multiline and requires greater height.