Bug 1405043 — Add Skip What's New and Skip Intro to all tests.
This tidies up the launchArguments to just being used in the TestAppDelegate, instead of spread around AppConstants. Each of SKIP_WHATSNEW, SKIP_INTRO and CLEAR_PROFILE can be used independently of one another.

Additionally, this cleans up the `restart()` method to the home button instead of `terminate()` and use `activate()` instead of `launch()`. This does not affect test setUps, but does help with tests that require restarts.