Bug 1401278 - Allow to long press suggestions. (#3208) r=farhan
* Bug 1223460 - Show search engine bar when returning to search mode. r=farhan

Right now, if you enter a search, load the search page, and then tap
the URL bar, the search terms are displayed but neither the search
engine bar nor the most recent search suggestions are displayed.  That
makes it awkward to redirect searches to an alternate search engine:
you need to tap the URL bar before the engine bar is displayed.

This patch fires urlBar(didEnterText:) when entering overlay mode and
the location is recognized as search terms.  That will re-query search
suggestions and display the engine bar.  This does not display the Top
Sites/Pocket Suggestions UI as frequently, but makes re-querying an
existing search work smoothly.

The existing code paths handle extracting search terms from
non-default engines already.

In the cases where the search term is a valid URL, no information is
leaked: the search term is recognized as a URL and no search
suggestion query is performed.

* Bug 1401278 - Allow to long press suggestions. r=farhan