Move connection acquisition to bg thread. Typing is liquid smooth now. `withConnection` calls `getSharedConnection` which can end up scheduled on the main thread. Because of the single queue of DB ops, the main thread will block while DB operations are running. On startup, this is adding multiple 1-sec main thread lockups. For this bug there are multiple db operations happening and blocking happens on the main thread.
Basically, the existing implementation is a workaround to fix background crashes.
Killing the performance of the entire app is not the way to workaround the crash.
SwiftData already has a flag for 'closed' that is set when the db is shutdown, if there is a bug with backgrounding still, we need a better solution than what was done here.