Add Archive pre-action call to scheme-postbuild script
Our test runner is trying to access a database for each test class that is stored in the `test-fixtures` directory of the  project. Our scheme defines a pre-action step for the Test action that copies the databases to the app bundle but it’s missing that step for the Archive action.