Adressing Bug 1488667 - Open Calendar Elements from downloads
This is a basic/quick update to adress the mentionned bug. This
simply checks if the mime type corresponds to a calendar once clicked
inside the *Downloads* sections and rely on internal iOS
`UIDocumentInteractionController` to open the file.

Too address a better solution for the user, aka opening without downloading,
this would require to download automatically the files and click without
passing by the download section and I have not enough knowledge about the application yet.