[focus-ios] Closes https://github.com/mozilla-mobile/focus-ios/issues/1546: Adds Swiftlint Git Pre-Commit Hook (https://github.com/mozilla-mobile/focus-ios/pull/1610)
* Adds Swiftlint Git Pre-Commit Hook

Adds a new bash script file which determines if swiftlint is installed
and runs swiftlint on the staged git files. First run performs an
autocorrect for simple fixes and then a second run of swiftlint
captures the exit status to determine if the commit should be
successful or not. README.md is updated for users to install swiftlint
as a requirement and the checkout script is updated to copy the new
bash script to a pre-commit hook.

Fixes: https://github.com/mozilla-mobile/focus-ios/issues/1546

* Test commit

* Test commit

* Adds the Strict Requirement For SwiftLint

Modifies the swiftlint command to add the strict option to treat
warnings as error messages.

* Remove lint test code