Bug 1523241 - TabsButton animations overlapping and showing wrong count (#4641)
* Bug 1523241 - TabsButton animations overlapping and showing wrong count
I am only seeing this on 12.1. This is the simplest fix without digging into
animation code. It is possible on 12.1 that colliding animations in this case result in
the later animation not completing (and thus not showing the final value of tab count).

* Add re-entrancy guard on tab updating function so that until the
animation is complete, it is not possible to trigger another animation.
The final value of the tab count at the end of the animation is guaranteed
to be the latest value.