Bug #5011 - Tab Tray inertia scroll w/ hiding find tabs bar (#5018)
The hiding of the tabs bar sets scrollView.contentInset, and setting
that breaks the inertial scroll (when finger lifted it stops immediately).
By setting this only on scrollDirection == .down, inertial scroll still
works when going up. It is still broken when scrolling down, the code as written
needs to set the contentInset in thi case.