Make auth failure during device registration drive FxA to "separated".
Previously, an auth failure during device registration would drive the
FxA state machine to "doghouse", which is a special terminal state for
when the FxA server tells the client that it needs to upgrade itself
before it can continue using the FxA APIs.

The appropriate state for auth failures is "separated", which prompts
the user to re-enter their password to reconnect to the account.