Webxr (#6614)
* wired up the "Stop AR" menu button

Now it sends a message into the Javascript context, in addition to stopping the session.

* remove duplicate code from merge

* (formatting)

* change external url handler to xrviewer

to not conflict with firefox for ios

* update browsing activity url string and spotlight search

* update INTERNAL_URL_SCHEME

* Update SiriShortcuts.swift

* Update AppConstants.swift

add new oauth client id
deal with variations of XRViewer scheme (mostly for dev)
disable leanplum

* update user agent

* updated apns push config to use "xvr"

* Update PushConfiguration.swift

* fixed push config

somewhere during the discussion, people started typing "xvr" instead of "xrv" ... should have been xrv.  fixed.

* small update of push config settings

fre small configu settings

* not sure if this UserActivityHander change is needed

I think it is (casing?)

* small local webserver that runs only in XRViewerDev

When built locally, XRViewer will try to start a local webserver on port 8080 in the app, IF AND ONLY IF you put a "Web" directory in the top level of the project.  The contents of that directory will be available via localhost:8080

* update cartfile

application-services needed to be updated to account for new swift version

* update to account for new api in moz dependency

* revert to old code, old XCode

* reverted Cartfile.resolved to old version