Stop generating `metrics.md` and direct to Glean Dictionary instead
Now that the Glean Dictionary (https://dictionary.telemetry.mozilla.org)
has hit production, our guidance is to move away from generating
metrics.md files inside individual applications. The Glean Dictionary
is a better solution for this task for a couple of reasons:

1. The Glean Dictionary contains a complete reference of all metrics
and pings sent by an application (not just those defined in the
application itself, as is the case with the markdown documentation)
2. The Glean Dictionary provides additional context and links beyond
what is provided by the documentation (for example, information on
how to access metric information via BigQuery or GLAM) as well as
annotations provided by data scientists in the course of analyzing data
(see, for example, the commentary section for:
https://dictionary.telemetry.mozilla.org/apps/firefox_ios/metrics/app_opened_as_default_browser).

This should also reduce some of the churn/burden on people when
they submit PRs, as there is no longer any need to update metrics.md
as a part of them.

Note that the data review guidelines have been updated to allow/reflect
the fact that the Glean Dictionary is the source of truth:

* https://github.com/mozilla/data-review/pull/27
* https://github.com/mozilla/data-review/pull/29