NavigationPath Fixes
NavigationPath.init was unable to cope with incoming URLs where the scheme was not all lowercase. As such, attempting to open a url such as "Http://www.apple.com" from the Notes app (or elsewhere) would fail when we are the Default Browser.

NavigationPath.init was too eager to handle 'normal' http[s] URLs as internal ones. i.e "http://glean.mywindows.com" would try to create a glean-based NavigationPath, instead of treating it as a web-url. Now all of our internal URLs have to have a scheme of "firefox" or "$(MozInternalURLScheme)" AND the (lowercased) host of that URL has be an exact match for what we are testing.

As part of this, a new build variable was added, $(MOZ_PUBLIC_URL_SCHEME), which is used instead of hard-coding "firefox" as our public scheme. Its use is analogous to $(MOZ_INTERNAL_URL_SCHEME). An extension to URL was added to read either the public or internal scheme and was used wherever the code used to check the plist itself.