[focus-ios] Fixes https://github.com/mozilla-mobile/focus-ios/pull/3128 - Bootstrap Taskgraph from pypi in decision task (https://github.com/mozilla-mobile/focus-ios/pull/3129)
* Support Python 3 in focusios_taskgraph

This in-lines the 'chunkify' package which doesn't support Python 3
and no longer exists on Github.

It was copied from:
https://hg.mozilla.org/mozilla-central/file/a8d37f94a6364103b7f373b9938ce7a62d9db9e6/python/mozbuild/mozbuild/chunkify.py

(where the Python 3 issue was already fixed)

* Fixes https://github.com/mozilla-mobile/focus-ios/pull/3128 - Bootstrap Taskgraph from pypi in decision task

Releng has tweaked the recommended way of bootstrapping Taskgraph from a
clone + checkout pinned revision, to Pypi and a requirements.txt.

This provides a few benefits:

1. Since Taskgraph uses SemVer, updating will provide a hint if there
   will be backwards incompatible changes.
2. Using version numbers is nicer than revisions (e.g, Taskgraph's
   CHANGELOG will be useful).
3. Provides a mechanism to pin other non-Taskgraph dependencies.

The reason I'm submitting this now is I'd like to migrate Taskgraph to a
new repository. By moving consumers over to Pypi first, the migration
can happen seamlessly without breaking anything.

Jira: RELENG-808

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>