FXIOS-4069 Soaking the nimbus features value() code
This PR introduces a function call `soakNimbus(featureId: in: at: )` which instantiates a `FeatureHolder` and calls it 10,0000 times.

We then call this everytime we need a `value()` from a feature.

Also in this PR: the experiment that is causing the bug to exhibit; the app should enroll itself at start up for developer builds.

Theory is that intermittent bugs become reliably reproducible if we can try the errant call lots of times.

Our assumptions here are:

1. The bug exists, but we need to find the right hardware/OS combination to exhibit it.
2. We've isolated the bug to the right call.