Bugfix FXIOS-6785 [v116] Page extent and content container (#15594)
* Fixing constraint issue when embedding content

The initial issue is that the homepage with bottom URL bar manages it's own status bar when scrolling. Since with coordinators we now have one and only container for both the webview and homepage, constrained needs to be adjusted to ensure the content appears as it should on screen.

Bottom constraint of the contentContainer is always the top of the overKeyboardContainer. Top constraint of the contentContainer was sometimes the bottom of the URL bar (header) or the bottom of the statusBarOverlay. This PR changes that so the top constraint of the contentContainer is always the bottom of the header stack view.

For this to work, the homepage cannot manage it's own status bar overlay anymore. There's a new StatusBarOverlay class introduced here, that will handle the status bar for us. Homepage has a delegate to tell when it scroll, so the status bar can be adjusted in the needed case of bottom URL bar on homepage. In all other cases, the status bar overlay is opaque.

* Fix comments

* Move back extension

* Fix tests

(cherry picked from commit 5cbdf5f0a51e3ecbf4e027c0f8fc88f968bcc663)

# Conflicts:
#	Client.xcodeproj/project.pbxproj
#	Client/Coordinators/Browser/BrowserCoordinator.swift
#	Client/Frontend/Home/HomepageViewController.swift
#	Tests/ClientTests/Frontend/ContentContainerTests.swift