temp-baseline ping: Send even if empty
The `active` ping (and probably the `dirty_startup`) ping will not have
any metrics outside the usual ping/client info, but we still want to
send it to get that signal.

This was missed in the initial implementation as I tested locally while
also setting another metric, which was ultimately removed before landing.
That makes the currently received data hard/impossible to analyse.