Updated PrimaryRoundedButton usage in FirefoxAccountSignInViewController
Using `withRenderingMode(.alwaysTemplate)` on the button's image so it gets tinted like the button's title when highlighted, interpolating from the base colour.
(This fixes a bug where the theme was not being properly taken into account in the signin page. If the button was highlighted when using dark mode, the image was white instead of dark)