Build FXIOS-8134 [v124] Stop setting dummy secrets for bitrise-trigger tasks
There's generally not any point in creating a dummy secret for pull
requests because the tasks won't do anything useful anyway. There's also
no security benefit of doing this because access to this secret is
already controlled in Firefox CI's configuration repo:
http://hg.mozilla.org/ci/ci-configuration/file/ae2d04f50811dad99be739092eadaf9effce3829/grants.yml#l1740

Instead, let's just always attempt to access the secret and focus on not
running these sensitive tasks where it isn't allowed in the first place.