Add FXIOS-8241 [v124] Fix icon and set .inactive for synced tabs (#18317)
Recent app-services added a new `inactive` boolean to tab payloads,
with the default value being false. This should be true if the
tab is shown in the browser as being one of the "inactive" tabs.

The intent is that desktop will start looking at this flag and treat
inactive tabs differently, offering an experience closer to what the
user sees re tabs on their mobile device.

Also started filling in the favicon for tabs when available. This
doesn't seem to handle all tabs - eg, wikipedia still has nil as the
favicon - but some tabs having an icon seems an incremental improvement
to zero tabs having an icon.