Add EXP-4268 [v125] Add action-params and change type of action
This is a breaking change, for experimenter users of the messaging feature.

It tightens up the type of `action` to use the string-alias `ActionName`, i.e. only keys of the `actions` table.

This is a large improvment in usability for message senders using the experimenter UI.

This removes the ability to open URLs directly in the action.

For this, the new `action-params` map is used to parameterize the action's deeplink URL.

The same string interpolation used on `action` is used on the `action` deeplink URL, and the values of the params.

For example, whereas before you may have said:

```json
"action": "https://example.com/survey?client={uuid}"
```

now you might write:

```json
"action": "OPEN_URL"
"action-params": {
    "private": "true",
    "url": "https://example.com/survey?client={uuid}"
}
```

Note that `action-params` are specific to each action, and so cannot be validated by the messsaging system.