Make `RustFirefoxSuggest` a class, not an actor.
Isolated methods on an actor are serialized, to protect the actor's
mutable state. This is undesired for `RustFirefoxSuggest`, because
it doesn't have any mutable state, and uses GCD queues to run read and
write operations concurrently and off-main-thread. Using an actor here
would serialize these operations, because isolated methods can't run
concurrently.

Instead, we can make `RustFirefoxSuggest` a regular class, and use
`withCheckedThrowingContinuation` to expose an `await`-compatible API.