Version bump restore 11.0.0 with CI fix (#875)
Some additions are added on top of the referenced PR in the commit.
This is a force-pushed commit to keep the latest version bump to 11.0.0
------
Added the possibility to strictly reference to `main` and `main-xxx` branches to spin up builds.
Major fixes are inherent to the `firefox-ios` path added to the `check_marketing_version.sh` script as well as some `yml` CI files for both GitHub Actions and CircleCI.
We are also referencing an empty Production file (git ignored) to make easy for the CI script to inject the required production-only configs.