Refactor Bugzilla-1959916 [Focus telemetry] Split out legacy metrics (#26079)
* Refactor Bug 1959916 - Split out legacy metrics

Telemetry is generally disabled, so these metrics are _not_ collected
anymore.
Due to the codebase still referencing them though we have to keep the
definitions around.
However to mark them as `removed` in the Dictionary we need to split
them out.

* Refactor Bug 1959916 - Remove the unused dau-reporting ping

This ping was replaced by the `usage-reporting` ping.
The Dictionary will soon correctly mark this as `removed`.