Add FXIOS-11506 [Relase Automation] Add required CI tasks to make releases through shipit (#26258)
* Add the scaffolding for release promotion and a promote phase

The scaffolding part is heavily inspired by the code from
firefox-android and nothing's special in there.

The promote workflow is basically `SPM_Deploy_Prod_Beta` with the
shipping to appstoreconnect part ripped out. It runs on bitrise through
a [small taskcluster <=> bitrise
interface](https://github.com/mozilla-releng/scriptworker-scripts/tree/master/bitrisescript)

The scheme decision ended up being messy but the idea is as follows:

- If the repo URL contains staging, override the wanted scheme to
  `FirefoxStaging` and export method to `development`
- Otherwise, assume it's a production build, use the `BITRISE_SCHEME`
  passed in the environment for the task and change `CODE_SIGN_IDENTITY`
  to distribution.

The `BITRISE_SCHEME` is always set to either `Firefox` or `FirefoxBeta`
by the promote task itself (but is ignored on staging builds, as per the
above). It is set to `FirefoxBeta` if the build is specifically for a
beta version and `Firefox` otherwise. A `beta` version is any version
that has a patch component in it, `139.0.0` is a beta version, `139.0`
isn't. The source of truth for this is the version specified in the
`version.txt` for the promoted commit.

Because taskgraph doesn't support `by-` keys in lists, I had to write
a transform that would replace a well known value
(`<release-type-scheme>`) in every specified environment variable for
the promotion tasks.

The promotion task will output 2 artifacts:

- A `Client.ipa` containing the app to upload to TestFlight
- A `symbols.zip` archive containing the associated debug symbols to
  upload to sentry

Those artifacts are uploaded to bitrise and reexported by the
taskcluster task. The second part is important for the next phase.

Co-Authored-By: Heitor Neiva <hneiva@mozilla.com>

* Add a push phase to release promotion

The push phase it fairly straightforwad. It grabs the artifacts from the
promote phase and pushes them where it needs to.

That is sentry for the debug symbols and TestFlight for the app itself.
We get those artifacts from the promote task through taskcluster as
they're already there and we can easily reference it back.

Co-Authored-By: Heitor Neiva <hneiva@mozilla.com>

* Add a ship phase to release promotion

This is by far the most taskcluster heave phase of release promotion. It
consists of 3 different things:

- First we create a github release/tag for the promoted commit through
  githubscript
- Second, we create a commit to bump the version to the next one through
  treescript
- Finally we mark the release as shipped in shipit, completing the
  release.

A few notable things:

- The github release will be marked as `prerelease` for beta versions, and
  as `latest` otherwise. This is so the "real" releases are always the
  ones showing on the github page for the project
- Version bumping will always conserve the "beta status" of the version.
  "139.0.0" will bump to "139.0.1". "139.0" will bump to "139.1". To
  switch from a release type to the other, a manual step is required
  (adding/removing the patch component from the version in
  `version.txt`)

Co-Authored-By: Heitor Neiva <hneiva@mozilla.com>

* Add missing license headers

* Solidify how we detect staging relpro by using the bitrise slug

We had 3 ways of detecting whether a promotion was running for staging
or not, this solidifies it so everything looks at the BITRISE_APP_SLUG
now

* nit: add a blank line in the bitrise YAML for readability

This is barely better tbh

* Use `set -xe` in all steps

Instead of setting them individually which just makes the file longer
for no reason. Also removed comments that aren't very useful

* Don't use dev workers

* Quote environment variables in relpro push workflow

* Fix the default value for workflows in relpro_promote transform

* Remove TODO

Staging repos used to not be a thing and people were using their own
forks and had to specially give access to automation for this t work.
That's not the case anymore

* Add a comment about the staging specific step

---------

Co-authored-by: Heitor Neiva <hneiva@mozilla.com>