Fix FXIOS-11506 [Release Automation] Use the bitrise app title to detect staging build promotions (#26404)
Use the bitrise app title to detect staging build promotions

In 6464e3c3f6833ffc4144ef6a773cc830c413d217 we misunderstood what the
app slug was and changed the way staging was detected to that quite late
in the process. Turns out the app slug is the unique identifier for the
app (a UUID), not the application name itself which means that those
`run_if` were never evaluated as `true`.