Fix FXIOS-11506 [Release Automation] Actually override `BITRISE_SCHEME` in promotion tasks (#26551)
Actually override `BITRISE_SCHEME` in promotion tasks

So it turns out that you cannot override default environment variables
through the API to trigger builds, bitrise will just ignore any
environment variable that has a default set. Because there's a global
default for `BITRISE_SCHEME` set to `Fennec`, it means that we're always
trying to build Fennec on promotion instead of Firefox or FirefoxBeta,
which of course, fails.

This uses a trick from [here](https://discuss.bitrise.io/t/how-to-override-environment-variables-through-the-build-trigger-api/191)
to override the default if there's a value to override it with by using
a different key and then forcing the value through envman. Note that we
were already forcing the value through envman for staging which is why
we didn't catch this earlier.

(cherry picked from commit 52dea71b169a688f6212c781329ca2550c5eb26b)