Fix [Release Automation] Don't try to add the Nightly group to Betas after a nightly is built (#27082)
Don't try to add the Nightly group to Betas after a nightly is built

In e528a9a4da9e7fe2f7ab203050cd84f0d263934f we tried adding a 2 minutes
wait between pushing the build to TestFlight and the addition of the
nightly group to said build. That is apparently not enough time for
apple to start returning that new build as the latest.

This partially reverts e528a9a4da9e7fe2f7ab203050cd84f0d263934f and
instead makes a `AddNightlyGroupToTestFlightBeta` workflow that we can
call called on a bitrise schedule long after the nightly finishes
building.