Replace references to staging in test params for taskgraph
This should fix the taskgraph tests on main. Technically only
`cron-beta-releases` needed that change as the test would get the list
of branches of the non staging repo but then would try to get the
version of that branch on the staging one. But there's no value in
keeping any of those params referencing staging so I just made them all
use the original repository.